fix issues with choosing namespaces
allow namespace selection to happen with options

can have multiple namespaces, could search each to find the
best one potentially. for now find the one that exists / has vectors.

fixed retries and handling so that we are much more robust and do
better error handling and waiting for data to the prompt.